+++ docs/CHANGE_LOG.md
@@ -16,6 +16,7 @@
 - Moved from ES2016 to ES2020.
 - Updated to the latest NPM packages.
 - Added an ignore path for ".DS_Store" files when packing the NuGet package.
+- Renamed some of the documentation.
 
 #### **Fixes:**
 - Fixed missing documentation in the "options/OPTIONS.md".